Create README.md
# Arduino Signal Generator

Arduino program that allows manipulation of the Si5351. It provides a set of fuctions via the LCD display.  A push button and a Rotary Encoder
controls the LCD menus.  

Its a work in progress there are a few improvements and a few bugs to be fixed. There is an issue with the Si5351 library I wrote 
for frequencies above 120 Mhz.  I need to troubleshoot and fix the problem

Dave, VE300I